requirements.txt
Django==5.2.4
djangorestframework==3.16.0
Pillow==11.3.0
python-decouple==3.8
django-cors-headers==4.3.1 